Add some docstrings and move some helpers to the utils file
